refactor: extract asset for url and implement rewrite
